Removed Source Code from Example Notebook
I removed the source code but left some of the explanations on how to use the function for users. The imports and functions run on my Jupyter Notebook now without the source code pasted. I hope the this has fixed the import errors. Please let me know if there is anything else needing updates.