Created 2nd Example Notebook Using the UVMOS
The second example notebook uses the uvmos point source function to generate a blackbody. There are two examples in this notebook, the first is for the star EV Lac during Quiescent Phase to create a Castor spectrum. The second is again EV Lac but during the Flaring phase. The main difference is temperature change. EV Lac will be observed by Castor when it launches. This notebook shows the differences in SNR for both cases to compare visually.

This commit also fixes a spelling mistake in the first example notebook.